stricter typing add integrity check to spi reference read

+++ rwchcd.h
@@ -52,8 +52,8 @@ enum {
 
 #define RWCHCD_SPI_MAX_TRIES	10	///< how many times SPI ops should be retried
 
-typedef unsigned long	temp_t;		// all temps are internally stored in Kelvin * 100 (long avoids overflow with disconnected sensors)
-typedef short		tempid_t;	// temperature index: if negative, is an offset. If > sizeof(Runtime->temps[]), invalid
+typedef int_fast32_t	temp_t;		// all temps are internally stored in Kelvin * 100 (32bit avoids overflow with disconnected sensors). Must be signed for maths
+typedef int_fast16_t	tempid_t;	// temperature index: if negative, is an offset. If > sizeof(Runtime->temps[]), invalid
 
 
 enum e_runmode {
@@ -79,7 +79,7 @@ enum e_systemmode {
 struct s_config {
 	bool configured;
 	time_t building_tau;		///< building time constant
-	short nsensors;			///< number of active sensors (== id of last sensor +1)
+	int_fast16_t nsensors;			///< number of active sensors (== id of last sensor +1)
 	tempid_t id_temp_outdoor;	///< outdoor temp
 	temp_t limit_tfrostmin;		///< outdoor temp for frost-protection
 	temp_t limit_tsummer;		///< outdoor temp for summer switch over
